Day 008 - Introduce DirectSound helper classes
I made some helper classes to manage the lifetimes of DirectSound resources. Initially, my code for DirectSound was panicking. I discovered that I was letting the `IDirectSound` object go out-of-scope. The way to prevent this in Rust is to use lifetimes.

I only figured out my bug yesterday, because I was confused. How could Windows be giving me back an object, but then I not need to explicitly de-allocate the object later? Since `IDirectSound` is on the stack, not a reference to something on the stack or a pointer to something on the heap, I was confused how polymorphism could work.

I vaguely remembered when learning about COM+ years ago that COM+ utilizes the [PIMPL pattern](https://en.cppreference.com/w/cpp/language/pimpl.html) to guarantee certain cross-platform guarantees. So the only thing on the stack is probably a pointer to the actual implementation and the destructor probably deallocates it for us upon going out-of-scope.

This commit is mostly me experimenting with how to capture those lifetimes in Rust. It certainly feels a bit heavy, but I also don't mind shrinking the size of `main.rs` again.