Day 009 - QA Session - Control sine wave with gamepad
There's this fascinating exercise in the Q/A session where Casey shows how to control the pitch of the sound using the gamepad. It helped me to identify some bugs I had in my sound code. For example, Casey switched to filling the buffer once upfront in the application rather than only in the loop. This way, if the write position is the same as the current play cursor, we can just write 0 bytes (a.k.a., do nothing). I do the same, now. I also have to jump through more hoops in Rust making sure arithmetic handles overflow/wrapping correctly in a few spots.