Day 011 - The basics of platform API design (WIP)
The idea of this episode (and the next) was to start separating out the Windows-specific behavior from the application logic. Basically, by the end of this exercise, all the Win32 behavior should be in one file (ignoring the direct sound stuff I did before), main should *just* be kicking off the code, and the platform-independent code should live in a separate file. I am not completely done - by the end of the episode, Casey basically had the platform-independent code kicking off the Win32 code, and I am still very much driving the application using Win32 code. There's a bit of two-way indirection involved with flipping everything on its head, so I'll focus on that in a follow-up commit. Mostly, I want to commit now to avoid losing anything.

I ran into a stupid bug because I passed a pointer to Windows and then moved the object pointed to. This is the sort of thing that makes Rust so appealing... this sort of crap is common in C/C++ but nearly impossible in Rust without unsafe code. I am curious if there's a way I could have designed my code to expose this bug before I made it. On a related note, I want to experiment with replacing `VirtualAlloc` et al. with a simple `Box<u8[]>` at some point, since it'd be one less Win32 operation.

On Day 012, Casey introduces a secondary buffer to avoid the complexity of dealing with a ring buffer in the platform-independent code. I am curious if I can avoid this insignificant overhead in a more interesting way, so I might try that out later.