Day 012 - Platform-independent sound output - refinements
I noticed Casey had the sound index in the platform-dependent layer, since it is how we are figuring out where in the DirectSound buffer to write, and it has nothing to do with generating the sine wave.

I also noticed Casey allocates the sound buffer exactly once. Since he knows that the buffer will never exceed the maximum length of the DirectSound buffer, he can initialize it once and just reuse it. In Rust, I emulate this using a mutable slice.