Day 018 - Enforcing a Video Frame Rate
We're basically trying to target a hard framerate, sleeping if we ever finish early. Casey was struggling with how to capture the actual refresh rate of the monitor, I think mostly because the documentation for `EnumDisplaySetting` was confusing. If I ever get back 0 or 1, I just return 60 as a default refresh rate.

I broke out the rendering logic so the application layer runs before sleeping, but the actual platform layer render happens after. I am curious if we'll end up doing something similar for sound. I sound was choppy at the end of the stream, so I suspect Casey will address that next time or as part of the Q/A. For now, I just made my refresh rate the same as the monitor refresh rate. My assumption is we'll end up realizing the sound buffer is configured for 48,000 samples per second, but we're writing to it every 3rd of a second or something and that's creating artifacts - not sure. It goes away when I match the screen refresh rate, though... weird.