Day 019 - Improving audio synchronization
In this episode, Casey did some neat visualizations of the sound play and write cursor. I didn't bother with the visualizations, although interesting. It looks like Casey is going to do a little more research into fixing the audio lag issue. I noticed yesterday that with 60fps, there was no audio lag - which Casey goes over a little bit in the Q/A session. It makes sense - with 60fps, the delay between frame limiting is 16ms instead of 33ms, so less lag is necessary. But perhaps 60fps is being too aggressive a target at this point. Curious to see what is done about it on day 20.