Day 020 - Debugging the audio sync
In this episode, Casey built a visualization tool into the rendering logic to help debug why the audio was skipping at the end of Day 019's episode. I did not bother with that exercise since it will eventually go away anyway.

I wrote a convenience method to retrieve both play and write cursors at the same time. Previously, we only used the play cursor, but now we need to both in order to determine whether the write cursor currently falls within the current frame. If it does, it means our sound card has low-latency, and we can write less to the buffer each time. If the latency is high, we write an extra frame's worth of audio. The big idea is we have to first measure a frame's worth of audio to see how far apart the play and write cursor are. Using that, we can check where the write cursor will land.

In the QA sessions, it was pointed out that the audio seems to change in hertz after waiting a while, which I had also noticed. Casey discovered it was simply a matter of resetting the variable to track the current sine wave position (what he called `tSine` and I call `sound_theta`).

My guess is the start of the next episode (or later on at least), there will be some adjustments to this code. Some of the logic isn't clicking for me, and experimenting showed results vary greatly depending on the safety bytes used. There's probably a bug hiding in here somewhere. My one monitor refreshes at 144hz vs 60hz, so I noticed issues sooner and experimented until the glitches went away.