Day 022 - Instantaneous live code editing
Casey doesn't want to wait for live updates to take effect. Instead, he wants to inspect the last modified timestamp on file to determine if a hot swap is necessary. His episode covers the oddities of debugging in VS and dealing with different PDB files. Intellij is a little different and I had to come up with a different strategy for swapping out files.

The cool thing is Rust hides a lot of the Windows API oddness and exposes file metadata (for the modified date/time) and grabbing the current EXE file name and parent directory. Rust also has `Path`/`PathBuf` classes that make building paths much easier/safer than what Casey was doing. I'm happy with how it turned out.