Day 023 - Looped live code editing (WIP)
Casey was trying to make a point about how important looped live code editing is to avoid having to constantly switch between the keyboard and controller when switching between editing and testing. He inadvertently piqued my interest in making a simple player that could move around on the screen. Then he made it jump, so obviously I needed to do that too, completely missing the point of this episode. I will circle back to the rest of the episode once I am done playing around.