Day 023 - Looped live code editing (WIP)
In the next part of the episode, Casey shows how easy it is to capture input in a file and then replay that input, by toggling the "L" key. In his case, he is working with simple C structs. My `InputState` struct, however, is using `Vec`, which cannot be simply serialized/deserialized since a `Vec` contain pointers. Instead, I did a little research for a basic binary serialization/deserialization library and found bincode. It was pretty easy, once I found the right documentation. I also made a little state machine for tracking the state of the recorder, switching between writing to the file and reading to the file, with buffering.