Day 023 - Looped live code editing (WIP)
Casey is moving both the input and the "game state" into his recorder. I am actually not sure where the code for processing input and updating the game state should live, but I get the impression that it should live in the plugin, so we can tweak the logic as necessary. This commit includes some minor improvements to the `PlaybackRecorder` and starts moving the input state and game state into one overarching struct. I think the next move is to move the logic for `handle_input` into the plugin and out of the game state itself. The game state ends up being a mindless blob of data.