Day 027 - Exploration-based architecture - Dumb contexts
I really didn't like how badly context objects behaved in Rust. In other languages, it's common to pass a single object around rather than a bunch of parameters, especially when crossing boundaries like the plugin API in this project. That way, if I need to add more "parameters" later, I only have to modify the context and any implementation using the new "parameter". If you didn't, you'd be changing the signature of the function, thus changing the API, and every implementation would need to be updated. This is usually a *really* bad thing when you're talking about plugins where there are several implementations. In our case, there are two implementations: the actual implementation and a thin wrapper that just redirects the actual implementation. Even with two, changing the API is annoying.

The reason context objects in Rust are annoying is due to the borrow checker. If one "parameter" is a mutable reference and another parameter is an immutable reference, the borrow checker doesn't allow mixing read-only operations with mutating operations directly, so I was making a bunch of annoying inline functions to avoid the overhead of redirection.

I thought about it for a while and realized I can just make the context short-lived, dumb structs. You construct them, pass them through, and immediately destruct them. Problem solved! This code looks so much better now.