Day 028 - Drawing a tile map - Render player with center of gravity at feet
Okay, I learned something. I was really confused. Whatever math I came up with to center the player at their feet (bottom middle), I couldn't figure out how to prevent the player from being able to walk off the screen. As soon as I got the limits working, changing the player size would break the calculation. That got me thinking that the issue had something to do with the requested window dimensions vs the actual client dimensions. It turned out the issue was somewhat related - when we requested a buffer to write into, we requested a buffer with the same dimensions as the window, not the client area. The call to `StretchDIBits` deals with shrinking the buffer dimensions into the actual client dimensions. I need to make sure any rendering logic I perform is in terms of the buffer dimensions, not the client dimensions. I played around with making the buffer the actual dimensions of the client area, but that ended up being quite a lot of effort and I never got it working.