Day 028 - Drawing a tile map - Avoid stretching the image
In my last commit, I was trying to figure out how to avoid stretching when using `StretchDIBits`. The client area is always going to be smaller than the requested window size, so the buffer always needed shrank. Now I grab the exact client area size and allocate a buffer to fit it perfectly. The piece I was missing before was that the actual buffer needed to correspond with the bitmap info's header data. Otherwise, the call reads outside the bounds of the buffer (oof).