Day 029 - Basic Tile Map Collision Checking - Helper method to detect collisions
Casey noticed that the screen flickers with the `PatBlt` call. I already noticed before and just removed it - but I liked his idea of just updating the offscreen buffer to black. He also suggested a helper method to detect if a single point was traversable or not, so I followed suit.