Day 031 - Tilemap Coordinate Systems - Experiment using units of measure
I want to see if we can use [uom](https://github.com/iliekturtles/uom) to represent lengths, like meters. Most of this commit is the result of bincode (for storing/restoring game state) not recognizing uom classes, such as `Length`. Fortunately, both libraries support serde, so we can update all of our previous uses of bincode directly with serde instead. And then only in the `PlaybackRecorder` deal with bincode.

Ideally, once we adopt uom, we can avoid a lot of nasty bugs related to invalid conversions. We might even be able to say how many pixels there are per meter, so conversions between pixels and meters will be safer. There have been a few other spots where Casey demonstrated converting between unit using dimensional analysis, so as we circle back around and get more familiar with uom, we might be able to adopt it in more and more places to avoid subtle bugs.