Day 033 - Virtualized Tile Maps - Track player as they walk around the world.
We no longer want to transition the entire screen once the player leaves the bounds a screen. Instead, we want to track the player, keeping them roughly in the middle of the screen, displaying all tiles on the left, right, top, and bottom within a certain range.

I spent a lot of time trying to figure out how to prevent displaying non-existent tile maps. I had some success, but the complexity really made it feel not worth it. If needed, I will probably revisit this in the future - I feel like the logic keeps getting simplified, so it might be easier some time when I haven't spent a weekend frying my brain.